Api helder

+++ app/Http/Controllers/api/TreballadorController.php
@@ -4,8 +4,9 @@ namespace App\Http\Controllers\api;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Models\Treballador;
 use App\Http\Resources\TreballadorResource as TreballadorResource;
-
+use Validator;
 
 class TreballadorController extends Controller
 {
@@ -20,28 +21,22 @@ class TreballadorController extends Controller
     //Mostrar fins a un màxim de 10 els treballadors de la farmàcia
     public function index()
     {
-        try {
-            //code...
+       
         
-        $Treballadors= Treballador::all();
-        $Treballadors= Treballador::Paginate(10);
+        $Treballador= Treballador::all();
+        $Treballador= Treballador::Paginate(10);
         
         $response = [
             'success' => true, 
             'message' => "Llista Treballadors recuperada",
-            'data' => $Treballadors, 
+            'data' => $Treballador, 
         ];
   
         return response()->json($response, 200);  
 
-    } catch (\Throwable $th) {
-        $response = [
-            'success' => false, 
-            'message' => "Error al buscar todos los trabajadores",
-            'data' => $th, 
-        ];
-        return response()->json($response, 404);  
-    }       
+    
+      
+         
     }
 
     /**
@@ -61,42 +56,42 @@ class TreballadorController extends Controller
     public function store(Request $request)
     {   
         
-        try {
-        
-        $testDni =  preg_match('/^[0-9]{8,8}[A-Z]$/g',$request->dni);
-        $validator = Validator::make($input, [
-             'nom' => 'required | min:3 | max:20' ,
-        ]);
-        if($validator->fails() || !$testDni){
-            $response = [
-                'success' => true, 
-                'message' => "errors de validacio",
-                'data' => $validator->errors()->all(), 
-            ];
-      
-            return response()->json($response, 404); 
-        }            
-        
-        $Treballador = new Treballador;
-        $Treballador->dni = $request->dni;
-        $Treballador->nom = $request->nom;
-        $Treballador->genere = $request->genere;
-        $Treballador->save();
-        $response = [
-            'success' => true, 
-            'message' => "Informacion Trabajador recuperada",
-            'data' => $Treballador, 
-        ];
-        return response()->json($response, 200);  
+        //
+        $input = $request->all();
 
-    } catch (\Throwable $th) {
+        // Creem un validador de les dades enviades, i li passem les regles
+        // que volem comprovar
+        $validator = Validator::make($input, [
+            'dni' => 'required|min:3',
+            'nom' => 'required|min:3',
+            'genere' => 'required|max:200'
+        ]);
+
+        // Si alguna dada no és correcta
+        if($validator->fails()){
+
+           $response = [
+             'success' => false,
+             'message' => "Alta incorrecta!",
+             'data' => $validator->errors()->all(),
+           ];
+           // Retornem l'array convertit a JSON i el codi d'error 404 de
+           //  HTTTP
+           return response()->json($response, 404);     
+        }
+
+                
+        $Treballador= Treballador::create($input);
+
+        // Responem a la crida amb un tot ok!
         $response = [
-            'success' => false, 
-            'message' => "Error al crear el trabajador",
-            'data' => $th, 
+           'success' => true,
+           'data'    => $Treballador,
+           'message' => "Alta correcta",
+
         ];
-        return response()->json($response, 404);  
-    }
+
+        return response()->json($response, 200);
     }
 
     /**
@@ -145,43 +140,65 @@ class TreballadorController extends Controller
         // $request->validate(
         //     [ 'name' => 'required | min:3 | max:20' ]
         // );+
-        $testDni =  preg_match('/^[0-9]{8,8}[A-Za-z]$/g',$request->dni);
-        $validator = Validator::make($input, [
-             'nom' => 'required | min:3 | max:20' ,
-        ]);
-        if($validator->fails() || !$testDni){
-            $response = [
-                'success' => true, 
-                'message' => "errors de validacio",
-                'data' => $validator->errors()->all(), 
-            ];
+    //     $testDni =  preg_match('/^[0-9]{8,8}[A-Za-z]$/g',$request->dni);
+    //     $validator = Validator::make($input, [
+    //          'nom' => 'required | min:3 | max:20' ,
+    //         //  'genere' => 'required | max:200'
+    //     ]);
+    //     if($validator->fails() || !$testDni){
+    //         $response = [
+    //             'success' => true, 
+    //             'message' => "errors de validacio",
+    //             'data' => $validator->errors()->all(), 
+    //         ];
       
-            return response()->json($response, 404); 
-        }
-        try {
-            //code...
+    //         return response()->json($response, 404); 
+    //     }
+    //     try {
+    //         //code...
 
-        $Treballador = Treballador::findOrFail($id);
+    //     $Treballador = Treballador::findOrFail($id);
+    //     $Treballador->dni = $request->dni;
+    //     $Treballador->nom = $request->nom;
+    //     $Treballador->genere = $request->genere;
+    //     $Treballador->save();
+        
+    //     $response = [
+    //         'success' => true, 
+    //         'message' => "Informacion Trabajador actualizada con exito",
+    //         'data' => $Treballador, 
+    //     ];
+  
+    //     return response()->json($response, 200); 
+    // } catch (\Throwable $th) {
+    //     $response = [
+    //         'success' => false, 
+    //         'message' => "Error al actualizar el trabajador",
+    //         'data' => $th, 
+    //     ];
+    //     return response()->json($response, 404);  
+    // }
+
+    $Treballador = Treballador::find($id);
+        if($Treballador==null) {
+            $response = [
+                'success' => false,
+                'message' => "Producte no trobat",
+                'data' => [],
+            ];        
+            return response()->json($response,404);
+        }   
         $Treballador->dni = $request->dni;
         $Treballador->nom = $request->nom;
-        $Treballador->genere = $request->genere;
+        $Treballador->genere = $request->genere;  
+
         $Treballador->save();
-        
         $response = [
-            'success' => true, 
-            'message' => "Informacion Trabajador actualizada con exito",
-            'data' => $Treballador, 
+            'success' => true,
+            'message' => "Treballador trobat",
+            'data' => $Treballador,
         ];
-  
-        return response()->json($response, 200); 
-    } catch (\Throwable $th) {
-        $response = [
-            'success' => false, 
-            'message' => "Error al actualizar el trabajador",
-            'data' => $th, 
-        ];
-        return response()->json($response, 404);  
-    }
+        return response()->json($response,200);
     }
 
     /**
@@ -194,24 +211,33 @@ class TreballadorController extends Controller
     //Eliminar els treballadors de la base de dades
     public function destroy($id)
     {
-        try {
-            //code...
-      
-        $Treballador = Treballador::findOrFail($id);
-        $Treballador->delete();
-        $response = [
-            'success' => true, 
-            'message' => "Trabajador eliminado"
-        ];
-  
-        return response()->json($response, 200); 
-    } catch (\Throwable $th) {
-        $response = [
-            'success' => false, 
-            'message' => "Error al eliminar el trabajador",
-            'data' => $th, 
-        ];
-        return response()->json($response, 404);  
-    }
+    // Eliminem el producte segons el codi que ens passin
+
+        // El busquem a la BD
+        $Treballador = Treballador::find($id);
+
+        // Si no trobem el producte responem amb informació
+        // sobre l'error
+        if($Treballador==null) {
+            $response = [
+            'success' => false,
+            'message' => "Producte no trobat",            
+            ];
+            return response()->json($response, 404); 
+
+        }
+        else { // El producte l'hem trobat
+
+            // posar dins try-catch en cas de haver-hi relacions clau forana!!
+            $Treballador->delete();
+
+            $response = [
+            'success' => true,
+            'data'    => $Treballador,
+            'message' => "Producte esborrat",
+            ];
+
+            return response()->json($response, 200);
+        }
     }
 }
